BA-2452-ba-content-feed-viewing-image-posts (#255)
As a user, on the Baseapp Content Feed,I would like to View the images
in a post that has images attached to it, In order to see images and any
additional text shared by users..

 
Acceptance Criteria 
Context
In this story we will implement the basic behaviors of a post that has
an image in it. From being able to view images to scrolling the
carrousel.

Business Rules

Given I am on the content feed page and a user has created a post with
images, when I scroll through the content feed, then I should see the
post with images displayed.

I should see the text content associated with the post as well.

Given I am viewing a post with multiple images in the content feed when
I view the post then the images should be displayed in a carousel
format.*

When I view the image carousel in the post, then the images should be
displayed with the following behavior.

I should see a "Previous" arrow button on the left side of the carousel

I should see a "Next" arrow button on the right side of the carousel

Clicking the "Next" arrow should display the next image in the order

Clicking the "Previous" arrow should display the previous image in the
order

The carousel should loop back to the first image after the last one is
displayed

I should be able to navigate through all images in the post using these
arrow buttons.

The image should be able to open in another tab, by right clicking and
selecting "open in another tab".

Design Link:
https://www.figma.com/design/XRD6wSl1m8Kz6XUcAy5CLp/BaseApp---WEB?node-id=2429-44307&node-type=frame&t=5TN6cVoKe9HL2Mvf-0

Approvd 
https://app.approvd.io/projects/BA/stories/35375 

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **New Features**
* Introduced a Content Feed Posts list with support for images and
reactions.
* Added components for creating, displaying, and interacting with posts,
including post creation forms, image carousels, and reaction buttons.
* Implemented responsive and virtualized post lists for improved
performance.
  * Added new icons for enhanced post interaction (reply, share, pill).
* **Enhancements**
* Improved type safety and structure for post-related data and component
props.
* **Removals**
* Removed legacy components and types related to the previous post
creation flow.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->